+++ discoveryapis_generator/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.9.16-dev
 
 - Require Dart `2.10` or greater.
-- Optional, named parameters are now sorted by name. 
+- Optional, named parameters are now sorted by name.
+- `generate` command: added `[--no]--delete-existing` flag. 
 
 ## v0.9.15
  - Add support for scopes defined in